app/requirements.txt
google-api-core==2.3.2
google-api-python-client==1.7.11
google-auth==2.7.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.5.2
google-cloud-bigquery==3.2.0
google-cloud-bigquery-storage==2.13.2
google-cloud-core==2.3.1
google-cloud-secret-manager==2.7.2
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.2
googleauthentication==0.0.16
grpc-google-iam-v1==0.12.4
SQLAlchemy==1.4.37
psycopg2
requests

